+++ src/backend/typeDefs/mutations/organization/obm/create.js
@@ -7,6 +7,22 @@ const createObmOrganizationTypeDefs = gql`
     organizationTiny: String
     businessModel: String
     start: Int
+
+    # "technology" fields
+    approvalTime: String
+    hasDecisionSupport: Boolean
+    hasPbMbAuthorization: Boolean
+    isEmrIntegrable: Boolean
+    medicalReview: String
+    treatmentSelection: String
+
+    # "vertical integration" fields that in the future should likely
+    # be handled by org-to-org connections
+    payer: String
+    pharmacyBenefitManager: String
+    specialtyPharmacy: String
+    labBenefitManager: String
+    parentCompany: String
   }
 
   type CreateObmOrganizationPayload {
@@ -16,6 +32,22 @@ const createObmOrganizationTypeDefs = gql`
     organizationTiny: String
     businessModel: String
     start: Int
+
+    # "technology" fields
+    approvalTime: String
+    hasDecisionSupport: Boolean
+    hasPbMbAuthorization: Boolean
+    isEmrIntegrable: Boolean
+    medicalReview: String
+    treatmentSelection: String
+
+    # "vertical integration" fields that in the future should likely
+    # be handled by org-to-org connections
+    payer: String
+    pharmacyBenefitManager: String
+    specialtyPharmacy: String
+    labBenefitManager: String
+    parentCompany: String
   }
 `