+++ src/backend/typeDefs/mutations/organization/obm/update.js
@@ -8,6 +8,22 @@ const updateObmOrganizationTypeDefs = gql`
     organizationTiny: String
     businessModel: String
     start: Int
+
+    # "technology" fields
+    approvalTime: String
+    hasDecisionSupport: Boolean
+    hasPbMbAuthorization: Boolean
+    isEmrIntegrable: Boolean
+    medicalReview: String
+    treatmentSelection: String
+
+    # "vertical integration" fields that in the future should likely
+    # be handled by org-to-org connections
+    payer: String
+    pharmacyBenefitManager: String
+    specialtyPharmacy: String
+    labBenefitManager: String
+    parentCompany: String
   }
 
   type UpdateObmOrganizationPayload {
@@ -17,6 +33,22 @@ const updateObmOrganizationTypeDefs = gql`
     organizationTiny: String
     businessModel: String
     start: Int
+
+    # "technology" fields
+    approvalTime: String
+    hasDecisionSupport: Boolean
+    hasPbMbAuthorization: Boolean
+    isEmrIntegrable: Boolean
+    medicalReview: String
+    treatmentSelection: String
+
+    # "vertical integration" fields that in the future should likely
+    # be handled by org-to-org connections
+    payer: String
+    pharmacyBenefitManager: String
+    specialtyPharmacy: String
+    labBenefitManager: String
+    parentCompany: String
   }
 `