+++ src/backend/typeDefs/queries.js
@@ -333,6 +333,7 @@ const queries = gql`
   }
 
   type ObmOrganization {
+    # base fields
     _id: ID!
     slug: String!
     type: String
@@ -340,6 +341,22 @@ const queries = gql`
     organizationTiny: String
     start: Int
     businessModel: String
+
+    # "technology" fields
+    approvalTime: String
+    hasDecisionSupport: Boolean
+    hasPbMbAuthorization: Boolean
+    isEmrIntegrable: Boolean
+    medicalReview: String
+    treatmentSelection: String
+
+    # "vertical integration" fields that in the future should likely
+    # be handled by org-to-org connections
+    payer: String
+    pharmacyBenefitManager: String
+    specialtyPharmacy: String
+    labBenefitManager: String
+    parentCompany: String
   }
 
   type ObmService {