+++ src/frontend/components/BusinessObjectModal/BusinessObjectModal/ButtonGroup/ButtonGroup.js
@@ -108,6 +108,8 @@ const ButtonGroup = ({
     )
   }
 
+  console.log(inputToUse)
+
   return (
     <div
       style={{